+++ src/redflag/pandas.py
@@ -20,9 +20,9 @@ limitations under the License.
 """
 import warnings
 
-from .imbalance import imbalance_degree, imbalance_ratio, minority_classes
+from .imbalance import imbalance_degree,minority_classes
 from .outliers import get_outliers
-from .target import is_continuous, dummy_classification_scores, dummy_regression_scores
+from .target import is_continuous, dummy_scores, dummy_regression_scores, dummy_classification_scores
 from .independence import is_correlated
 
 
@@ -72,18 +72,8 @@ class SeriesAccessor:
             warnings.warn('The Series does not seem categorical.')
         return minority_classes(self._obj)
 
-    def dummy_scores(self, task=None, random_state=None):
-        if task is None:
-            task = 'regression' if is_continuous(self._obj) else 'classification'
-
-        if task == 'classification':
-            scores = dummy_classification_scores(self._obj, random_state=random_state)
-        elif task == 'regression':
-            scores = dummy_regression_scores(self._obj)
-        else:
-            raise ValueError("`task` must be 'classification' or 'regression', or None to decide automatically.")
-
-        return scores
+    def dummy_scores(self, task='auto', random_state=None):
+        return dummy_scores(self._obj, task=task, random_state=random_state)
 
     def report(self, random_state=None):
         results = {}